refactor: added todo plan to incorporate validation of unique phases for update
