+++ tests/test_schema_post_bid.py
@@ -105,3 +105,71 @@ def test_validate_feedback():
 
     with pytest.raises(ValidationError):
         PostBidSchema().load(data)
+
+
+# Case 7: Neither success nor failed fields phase can be more than 2
+def test_phase_greater_than_2():
+    data = {
+        "tender": "Business Intelligence and Data Warehousing",
+        "client": "Office for National Statistics",
+        "bid_date": "2023-06-21",
+        "alias": "ONS",
+        "bid_folder_url": "https://organisation.sharepoint.com/Docs/dummyfolder",
+        "feedback": {
+            "description": "Feedback from client in detail",
+            "url": "https://organisation.sharepoint.com/Docs/dummyfolder/feedback",
+        },
+        "success": [{"phase": 1, "has_score": True, "out_of": 36, "score": 30}],
+        "failed": {"phase": 3, "has_score": True, "score": 20, "out_of": 36},
+    }
+
+    with pytest.raises(ValidationError, match="Must be one of: 1, 2."):
+        PostBidSchema().load(data, partial=True)
+
+
+# Case 8: Success cannot have the same phase in the list
+def test_phase_already_exists_in_success():
+    data = {
+        "tender": "Business Intelligence and Data Warehousing",
+        "client": "Office for National Statistics",
+        "bid_date": "2023-06-21",
+        "alias": "ONS",
+        "bid_folder_url": "https://organisation.sharepoint.com/Docs/dummyfolder",
+        "feedback": {
+            "description": "Feedback from client in detail",
+            "url": "https://organisation.sharepoint.com/Docs/dummyfolder/feedback",
+        },
+        "success": [
+            {"phase": 1, "has_score": True, "out_of": 36, "score": 30},
+            {"phase": 1, "has_score": True, "out_of": 50, "score": 60},
+        ],
+    }
+
+    with pytest.raises(
+        ValidationError,
+        match="Phase values must be unique",
+    ):
+        PostBidSchema().load(data, partial=True)
+
+
+# Case 9: Success cannot contain same phase value as failed
+def test_same_phase():
+    data = {
+        "tender": "Business Intelligence and Data Warehousing",
+        "client": "Office for National Statistics",
+        "bid_date": "2023-06-21",
+        "alias": "ONS",
+        "bid_folder_url": "https://organisation.sharepoint.com/Docs/dummyfolder",
+        "feedback": {
+            "description": "Feedback from client in detail",
+            "url": "https://organisation.sharepoint.com/Docs/dummyfolder/feedback",
+        },
+        "failed": {"phase": 1, "has_score": True, "score": 20, "out_of": 36},
+        "success": [{"phase": 1, "has_score": True, "out_of": 36, "score": 30}],
+    }
+
+    with pytest.raises(
+        ValidationError,
+        match="Phase values must be unique",
+    ):
+        PostBidSchema().load(data, partial=True)
